Improved HUD
It’s not really an HUD, but I’m calling it an HUD and it now includes
XP and ability cooldowns (even though abilities don’t work yet). Also,
I created a new file (because I dislike scrolling and I had to stick a
bunch of getter methods on the unit class).